+++ benches/benchmarks/iter_with_large_setup.rs
@@ -1,32 +1,26 @@
-#![allow(deprecated)]
-
-use criterion::{criterion_group, Benchmark, Criterion, Throughput};
+use criterion::{criterion_group, Criterion, Throughput};
 use std::time::Duration;
 
 const SIZE: usize = 1024 * 1024;
 
 fn large_setup(c: &mut Criterion) {
-    c.bench(
-        "iter_with_large_setup",
-        Benchmark::new("large_setup", |b| {
-            // NOTE: iter_with_large_setup is deprecated. Use iter_batched instead.
-            b.iter_with_large_setup(
-                || (0..SIZE).map(|i| i as u8).collect::<Vec<_>>(),
-                |v| v.clone(),
-            )
-        })
-        .throughput(Throughput::Bytes(SIZE as u64)),
-    );
+    let mut group = c.benchmark_group("iter_with_large_setup");
+    group.throughput(Throughput::Bytes(SIZE as u64));
+    group.bench_function("large_setup", |b| {
+        // NOTE: iter_with_large_setup is deprecated. Use iter_batched instead.
+        b.iter_with_large_setup(
+            || (0..SIZE).map(|i| i as u8).collect::<Vec<_>>(),
+            |v| v.clone(),
+        )
+    });
 }
 
 fn small_setup(c: &mut Criterion) {
-    c.bench(
-        "iter_with_large_setup",
-        Benchmark::new("small_setup", |b| {
-            // NOTE: iter_with_large_setup is deprecated. Use iter_batched instead.
-            b.iter_with_large_setup(|| SIZE, |size| size)
-        }),
-    );
+    let mut group = c.benchmark_group("iter_with_large_setup");
+    group.bench_function("small_setup", |b| {
+        // NOTE: iter_with_large_setup is deprecated. Use iter_batched instead.
+        b.iter_with_large_setup(|| SIZE, |size| size)
+    });
 }
 
 fn short_warmup() -> Criterion {