+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - IN DEVELOPMENT
+
+### Removed
+- The `Criterion::can_plot` function has been removed.
+- The `Criterion::bench_function_over_inputs` function has been removed.
+- The `Criterion::bench_functions` function has been removed.
+- The `Criterion::bench` function has been removed.
+
 ## [Unreleased]
 ### Fixed
 - Corrected `Criterion.toml` in the book.